requirements.txt
fastapi==0.116.1
Pillow==11.3.0
transformers==4.55.4
